requirements.txt
django>=1.5.5
django-dingos>=0.1.0
django-mantis-core=>0.1.0

# Additional requirements go here